Renamed characters namespace
