+++ Assets/Scripts/Characters/Adventurer.cs.meta
